framework/aifw, examples/aifw_test: Resolve review comments
1. framework/src/aifw/Kconfig: Resolve warnings
2. framework/src/aifw/TFLM.cpp: Change data type of variable storing
output of ftell() from int to long.
3. framework/aifw: Replace #ifndef include guards with #pragma once
4. examples/aifw_test: Apply TizenRT coding style

+++ framework/src/aifw/TFLM.cpp
@@ -120,18 +120,18 @@ AIFW_RESULT TFLM::loadModel(const char *file)
 		return AIFW_ERROR_FILE_ACCESS;
 	}
 	fseek(fp, 0, SEEK_END);
-	int size = ftell(fp);
+	long size = ftell(fp);
 	fseek(fp, 0, SEEK_SET);
 	if (size < 0) {
 		fclose(fp);
-		AIFW_LOGE("File %s size read as %d is invalid, errno %d", file, size, errno);
+		AIFW_LOGE("File %s size read as %ld is invalid, errno %d", file, size, errno);
 		return AIFW_ERROR_FILE_ACCESS;
 	}
-	AIFW_LOGV("Model File Size: %d", size);
+	AIFW_LOGV("Model File Size: %ld", size);
 	this->mBuf = (char *)malloc(size);
 	if (!this->mBuf) {
 		fclose(fp);
-		AIFW_LOGE("Memory not enough to allocate %d", size);
+		AIFW_LOGE("Memory not enough to allocate %ld", size);
 		return AIFW_NO_MEM;
 	}
 	fread(this->mBuf, 1, size, fp);
